Fixed 2 related bugs when connecting to EPICS channels.
In EpicsBaseImpl, if a binding attempt fails, then the next one would return OK even if it is not.
In ChannelBindingSupport, a connection always succeeded, because you are supposed to use the listener to check if it is OK. The problem is we use the same CAJContext for EpicsBaseImpl, so when you then try to connect there, it succeeds because the other one seems to be connected. Now ChannelBindingSupport will fail if the new channel is not available.

git-svn-id: http://source.gemini.edu/software/giapi/giapi-osgi/trunk@36655 ee6ba543-07f8-4773-b02c-01d50e9d00ba